move codec logic into tablecodec package
